+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file tracks changes to this project. It follows the [Keep a Changelog format](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+This version removes all external dependencies of `go-sse`. All our bugs are belong to us!
+
+### Changed
+
+- `Server.Logger` is now of a new type: the `Logger` interface. The dependency on x/exp/slog is removed. This opens up the possibility to adapt any existing logger to be usable with `Server`.
+
+### Added
+
+- The `Logger` interface, `LogLevel` type, and `LogLevel(Info|Warn|Error)` values.
+
 ## [0.7.0] - 2023-11-19
 
 This version overhauls connection retry and fixes the connection event dispatch order issue. Some internal changes to Joe were also made, which makes it faster and more resilient.